Some more changes for PR 3
