+++ src/utils/GeoUtils.ts
@@ -1,3 +1,7 @@
+import {PlaneAlert} from "../index";
+import {Aircraft} from "../models/Aircraft";
+import {AircraftMeta} from "../models/Airline";
+
 export class GeoUtils {
     public static distanceBetweenCoordinates(lat1: number, lon1: number, lat2: number, lon2: number) {
         let R = 6371; // km
@@ -14,6 +18,37 @@ export class GeoUtils {
 
     private static toRad(value: number): number {
         return value * Math.PI / 180;
+    }
+
+    public static findNearestAirport(aircraft: Aircraft | AircraftMeta) {
+        if (aircraft.meta.lon === null || aircraft.meta.lat === null || PlaneAlert.airports === null) {
+            return null;
+        }
+        PlaneAlert.log.debug(`Plane ${this.name} (${aircraft.icao}) searching for nearest airport of ${aircraft.meta.lat}/${aircraft.meta.lon}`);
+        let min_distance = Number.MAX_SAFE_INTEGER;
+        let nearest_airport = null;
+        for (const airport of PlaneAlert.airports) {
+            if (airport.type === 'closed') {
+                continue;
+            }
+            if (aircraft instanceof Aircraft) {
+                // We don't have supported airport types for aircraft we don't know beforehand
+                //Maybe in the future this can come form the API itself
+                if (aircraft.allowedAirports.indexOf(airport.type) === -1) {
+                    continue;
+                }
+            }
+
+            const distance = GeoUtils.distanceBetweenCoordinates(aircraft.meta.lat, aircraft.meta.lon, airport.latitude_deg, airport.longitude_deg);
+            if (distance < min_distance) {
+                min_distance = distance;
+                nearest_airport = airport;
+            }
+        }
+        return {
+            airport: nearest_airport,
+            distance: min_distance
+        };
     }
 
 }
